requirements/base.txt
Pillow
Django==2.2.5
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
django-extensions==2.2.1
django-environ
djangorestframework==3.10.3
django-model-utils==3.2.0
django-imagekit
drf-yasg==1.16.1
django-redis
djangorestframework_simplejwt
uuid